requirements.txt
flask==3.0.3
git-filter-repo==2.45.0
grpcio==1.65.5
importlib-metadata==8.4.0
importlib-resources==6.4.0
ipykernel==6.29.5
jaraco.text==3.12.1
lz4==4.3.3
openai==1.42.0
openai-whisper==20231117
ordered-set==4.1.0
pickleshare==0.7.5
pinecone-client==5.0.1
pip-chill==1.0.3
protoc-gen-openapiv2==0.0.1
python-dotenv==1.0.1
tomli==2.0.1
